requirements.txt
jira
PyYAML
plotly
stashy
python-dateutil
numpy
